progress

+++ pose_estimation/pose_data_npz.py
@@ -262,4 +262,4 @@ class PoseDataNPZSegmentationTorch(torch.utils.data.Dataset):
 
         color = np.transpose(color, (2, 0, 1))
 
-        return color.astype(np.float32)/255, label.astype(np.long), np.array(key)
+        return color.astype(np.float32)/255, label, np.array(key)
